Design:큐레이션 상세 페이지에서 지도 맵 여는 버튼 디자인 수정
