--- app/api/scrap/state/[id]/route.ts
@@ -1,24 +1,0 @@
-import { getSession } from "@common/utils/session/getSession";
-import { NextResponse } from "next/server";
-
-export async function GET(
-  request: Request,
-  { params }: { params: { id: number } }
-) {
-  const auth_info = await getSession();
-  const token = auth_info?.data?.accessToken;
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_SERVER_API}/api/v1/spaces/scraps/${params.id}`,
-    {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${token}`,
-      },
-      next: { tags: ["getScrapStateById"] },
-      cache: "force-cache",
-    }
-  );
-  const data = await res.json();
-
-  return NextResponse.json(data);
-}
