+++ app/feature/curation/components/CurationTopAppBar/curationTopAppBar.tsx
@@ -7,8 +7,6 @@ import MenuIcon from "@common/assets/icons/menu/MenuIcon";
 import { useEffect, useState } from "react";
 import CurationMenuModal from "../CurationModal/CurationMenuModal";
 import { copyLink } from "@common/utils/text/copy-link";
-import MapIcon from "@common/assets/icons/map/map";
-import Map from "@feature/map/components/Map/Map";
 import { CurationDetailResponse } from "@feature/curation/queries/dto/curation-detail";
 import BasicTopBar from "@common/components/ui/topBar/BasicTopBar/BasicTopBar";
 import { sliceText } from "@common/utils/text/slice-text";
@@ -31,19 +29,9 @@ export default function CurationTopAppBar({
   className,
 }: CurationTopAppBarProps) {
   const [menuModalOpen, setMenuModalOpen] = useState(false);
-  const [mapOpen, setMapOpen] = useState(false);
   const [linkCopyToastOpen, setLinkCopyToastOpen] = useState(false);
   const [toastText, setToastText] = useState("");
   const pathname = usePathname();
-  const [mapPlaceData, setMapPlaceData] = useState<
-    {
-      address: string;
-      name: string;
-      type: string;
-      purpose: string[];
-      imgUrl: string;
-    }[]
-  >([]);
 
   const handleMenuClick = () => {
     setMenuModalOpen(true);
@@ -53,10 +41,6 @@ export default function CurationTopAppBar({
     copyLink(pathname);
     setLinkCopyToastOpen(true);
     setToastText("링크가 복사되었어요");
-  };
-
-  const handleMapClick = (state: boolean) => {
-    setMapOpen(state);
   };
 
   const handleScrapDelete = async () => {
@@ -90,20 +74,6 @@ export default function CurationTopAppBar({
   };
 
   useEffect(() => {
-    setMapPlaceData(
-      curationDetail.spaceDetails.map((space) => {
-        return {
-          address: space.address,
-          name: space.name,
-          type: space.type,
-          purpose: space.purpose.split(","),
-          imgUrl: space.imageUrls[0],
-        };
-      })
-    );
-  }, [curationDetail]);
-
-  useEffect(() => {
     let timeoutId: NodeJS.Timeout;
     if (linkCopyToastOpen) {
       timeoutId = setTimeout(() => {
@@ -120,10 +90,6 @@ export default function CurationTopAppBar({
         <div className="w-[100%] flex items-center justify-between gap-[0.8rem]">
           <h1 className="headline3-semibold">{text && sliceText(text, 16)}</h1>
           <div className="flex items-center gap-[0.6rem]">
-            <MapIcon
-              className="mr-[0.4rem]"
-              onClick={() => handleMapClick(true)}
-            />
             {variant === "others" ? (
               <>
                 {curationDetail.isScraped ? (
@@ -142,14 +108,6 @@ export default function CurationTopAppBar({
           </div>
         </div>
       </BasicTopBar>
-      {mapOpen && (
-        <Map
-          placeData={mapPlaceData}
-          zoom={13}
-          handleMapOpen={handleMapClick}
-          className="fixed top-[7rem] z-10"
-        />
-      )}
       {menuModalOpen && (
         <div className="w-[100%] h-[100%] fixed top-0 left-0 z-50">
           <CurationMenuModal
